rename horizon_api file to be clear it holds types
