+++ src/index.ts
@@ -4,7 +4,7 @@ require("es6-promise").polyfill();
 import { version } from "../package.json";
 
 // Expose all types
-export * from "./horizon_api";
+export * from "./horizon_api_types";
 export * from "./server_types";
 
 // stellar-sdk classes to expose
